@@ -17,25 +17,32 @@
     ],
     "provide": {
         "psr/cache-implementation": "1.0",
-        "psr/simple-cache-implementation": "1.0"
+        "psr/simple-cache-implementation": "1.0",
+        "symfony/cache-implementation": "1.0"
     },
     "require": {
-        "php": ">=5.5.9",
+        "php": "^7.1.3",
         "psr/cache": "~1.0",
         "psr/log": "~1.0",
-        "psr/simple-cache": "^1.0"
+        "symfony/cache-contracts": "^1.1.7|^2",
+        "symfony/service-contracts": "^1.1|^2",
+        "symfony/var-exporter": "^4.2|^5.0"
     },
     "require-dev": {
         "cache/integration-tests": "dev-master",
         "doctrine/cache": "~1.6",
-        "doctrine/dbal": "~2.4",
-        "predis/predis": "~1.0"
+        "doctrine/dbal": "~2.5",
+        "predis/predis": "~1.1",
+        "psr/simple-cache": "^1.0",
+        "symfony/config": "^4.2|^5.0",
+        "symfony/dependency-injection": "^3.4|^4.1|^5.0",
+        "symfony/var-dumper": "^4.4|^5.0"
     },
     "conflict": {
-        "symfony/var-dumper": "<3.3"
-    },
-    "suggest": {
-        "symfony/polyfill-apcu": "For using ApcuAdapter on HHVM"
+        "doctrine/dbal": "<2.5",
+        "symfony/dependency-injection": "<3.4",
+        "symfony/http-kernel": "<4.4",
+        "symfony/var-dumper": "<4.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Cache\\": "" },
@@ -46,7 +53,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "3.3-dev"
+            "dev-master": "4.4-dev"
         }
     }
 }
