@@ -9,79 +9,118 @@ use SebastianBergmann\Git\Git;
 
 class GitSync extends Git
 {
-    private $user;
-    private $password;
-    protected $grav;
-    protected $config;
-    protected $repositoryPath;
-    static public $instance = null;
+    /** @var static */
+    static public $instance;
 
-    public function __construct(Plugin $plugin = null)
+    /** @var Grav */
+    protected $grav;
+    /** @var Plugin */
+    protected $plugin;
+    /** @var array */
+    protected $config;
+    /** @var string */
+    protected $repositoryPath;
+
+    /** @var string|null */
+    private $user;
+    /** @var string|null */
+    private $password;
+
+    public function __construct()
     {
         $this->grav = Grav::instance();
         $this->config = $this->grav['config']->get('plugins.git-sync');
         $this->repositoryPath = isset($this->config['local_repository']) && $this->config['local_repository'] ? $this->config['local_repository'] : USER_DIR;
+
         parent::__construct($this->repositoryPath);
+
         static::$instance = $this;
 
-        $this->user = isset($this->config['user']) ? $this->config['user'] : null;
-        $this->password = isset($this->config['password']) ? $this->config['password'] : null;
+        $this->user = $this->config['user'] ?? null;
+        $this->password = $this->config['password'] ?? null;
 
-        unset($this->config['user']);
-        unset($this->config['password']);
+        unset($this->config['user'], $this->config['password']);
     }
 
-    static public function instance()
+    /**
+     * @return static
+     */
+    public static function instance()
     {
-        return static::$instance = is_null(static::$instance) ? new static : static::$instance;
+        if (null === static::$instance) {
+            static::$instance = new static;
+        }
+
+        return static::$instance;
     }
 
+    /**
+     * @return string|null
+     */
     public function getUser()
     {
         return $this->user;
     }
 
+    /**
+     * @return string|null
+     */
     public function getPassword()
     {
         return $this->password;
     }
 
-    public function setConfig($obj)
+    /**
+     * @param array $config
+     */
+    public function setConfig($config)
     {
-        $this->config = $obj;
+        $this->config = $config;
         $this->user = $this->config['user'];
         $this->password = $this->config['password'];
     }
 
+    /**
+     * @return array
+     */
     public function getRuntimeInformation()
     {
-        $result = array(
+        $result = [
             'repositoryPath' => $this->repositoryPath,
             'username' => $this->user,
             'password' => $this->password
-        );
+        ];
+
         foreach ($this->config as $key => $item) {
             if (is_array($item)) {
                 $count = count($item);
                 $arr = $item;
-                if ($count == 0) {// empty array, could still be associative
+                if ($count === 0) {// empty array, could still be associative
                     $arr = '[]';
                 } else if (isset($item[0])) {// fast check for plain array with numeric keys
-                    $arr = '[\'' . join('\', \'', $item) . '\']';
+                    $arr = '[\'' . implode('\', \'', $item) . '\']';
                 }
                 $result[$key] = $arr;
             } else {
                 $result[$key] = $item;
             }
         }
+
         return $result;
     }
 
+    /**
+     * @param string $url
+     * @return string[]
+     */
     public function testRepository($url)
     {
-        return $this->execute("ls-remote \"${url}\"");
+        return $this->execute("ls-remote \"{$url}\"");
     }
 
+    /**
+     * @return bool
+     */
     public function initializeRepository()
     {
         if (!Helper::isGitInitialized()) {
@@ -96,6 +135,11 @@ class GitSync extends Git
         return true;
     }
 
+    /**
+     * @param string|null $name
+     * @param string|null $email
+     * @return bool
+     */
     public function setUser($name = null, $email = null)
     {
         $name = $this->getConfig('git', $name)['name'];
@@ -107,11 +151,16 @@ class GitSync extends Git
         return true;
     }
 
+    /**
+     * @param string|null $name
+     * @return bool
+     */
     public function hasRemote($name = null)
     {
         $name = $this->getRemote('name', $name);
 
         try {
+            /** @var string $version */
             $version = Helper::isGitInstalled(true);
             // remote get-url 'name' supported from 2.7.0 and above
             if (version_compare($version, '2.7.0', '>=')) {
@@ -131,7 +180,7 @@ class GitSync extends Git
     public function enableSparseCheckout()
     {
         $folders = $this->config['folders'];
-        $this->execute("config core.sparsecheckout true");
+        $this->execute('config core.sparsecheckout true');
 
         $sparse = [];
         foreach ($folders as $folder) {
@@ -154,7 +203,7 @@ class GitSync extends Git
                 foreach ($subfolders as $index => $subfolder) {
                     $last = $index === (count($subfolders) - 1);
                     $nested_tracking .= $subfolder . '/';
-                    if (!in_array('!/' . $nested_tracking, $ignore)) {
+                    if (!in_array('!/' . $nested_tracking, $ignore, true)) {
                         $ignore[] = rtrim($nested_tracking . (!$last ? '*' : ''), '/');
                         $ignore[] = rtrim('!/' . $nested_tracking, '/');
                     }
@@ -169,24 +218,34 @@ class GitSync extends Git
         $file->free();
     }
 
+    /**
+     * @param string|null $alias
+     * @param string|null $url
+     * @param bool $authenticated
+     * @return string[]
+     */
     public function addRemote($alias = null, $url = null, $authenticated = false)
     {
         $alias = $this->getRemote('name', $alias);
         $url = $this->getConfig('repository', $url);
 
         if ($authenticated) {
-            $user = $this->user ?: $this->config->get('user');
-            $password = Helper::decrypt($this->password ?: $this->config->get('password'));
+            $user = $this->user ?? '';
+            $password = $this->password ? Helper::decrypt($this->password) : '';
             $url = Helper::prepareRepository($user, $password, $url);
         }
 
         $command = $this->hasRemote($alias) ? 'set-url' : 'add';
 
-        return $this->execute("remote ${command} ${alias} \"${url}\"");
+        return $this->execute("remote {$command} {$alias} \"{$url}\"");
     }
 
+    /**
+     * @return string[]
+     */
     public function add()
     {
+        /** @var string $version */
         $version = Helper::isGitInstalled(true);
         $add = 'add';
 
@@ -212,6 +271,10 @@ class GitSync extends Git
         return $this->execute($add . ' ' . implode(' ', $paths));
     }
 
+    /**
+     * @param string $message
+     * @return string[]
+     */
     public function commit($message = '(Grav GitSync) Automatic Commit')
     {
         $authorType = $this->getGitConfig('author', 'gituser');
@@ -219,21 +282,21 @@ class GitSync extends Git
             $authorType = 'gituser';
         }
 
-        // @TODO: After 1.6 it should be changed to `$configMessage ?? $message`
         // get message from config, it any, or stick to the default one
         $config = $this->getConfig('git', null);
-        $message = isset($config['message']) ? $config['message'] : $message;
+        $message = $config['message'] ?? $message;
 
         // get Page Title and Route from Post
         $uri = $this->grav['uri'];
         $page_title = $uri->post('data.header.title');
         $page_route = $uri->post('data.route');
 
-        $pageTitle = $page_title ? $page_title : 'NO TITLE FOUND';
-        $pageRoute = $page_route ? $page_route : 'NO ROUTE FOUND';
+        $pageTitle = $page_title ?: 'NO TITLE FOUND';
+        $pageRoute = $page_route ?: 'NO ROUTE FOUND';
 
         // include page title and route in the message, if placeholders exist
         $message = str_replace('{{pageTitle}}', $pageTitle, $message);
+        /** @var string $message */
         $message = str_replace('{{pageRoute}}', $pageRoute, $message);
 
         switch ($authorType) {
@@ -260,9 +323,15 @@ class GitSync extends Git
         $author = '--author="' . $author . '"';
         $message .= ' from ' . $user;
         $this->add();
-        return $this->execute("commit " . $author . " -m " . escapeshellarg($message));
+
+        return $this->execute('commit ' . $author . ' -m ' . escapeshellarg($message));
     }
 
+    /**
+     * @param string|null $name
+     * @param string|null $branch
+     * @return string[]
+     */
     public function fetch($name = null, $branch = null)
     {
         $name = $this->getRemote('name', $name);
@@ -271,11 +340,17 @@ class GitSync extends Git
         return $this->execute("fetch {$name} {$branch}");
     }
 
+    /**
+     * @param string|null $name
+     * @param string|null $branch
+     * @return string[]
+     */
     public function pull($name = null, $branch = null)
     {
         $name = $this->getRemote('name', $name);
         $branch = $this->getRemote('branch', $branch);
-        $version = $version = Helper::isGitInstalled(true);
+        /** @var string $version */
+        $version = Helper::isGitInstalled(true);
         $unrelated_histories = '--allow-unrelated-histories';
 
         // --allow-unrelated-histories starts at 2.9.0
@@ -286,6 +361,11 @@ class GitSync extends Git
         return $this->execute("pull {$unrelated_histories} -X theirs {$name} {$branch}");
     }
 
+    /**
+     * @param string|null $name
+     * @param string|null $branch
+     * @return string[]
+     */
     public function push($name = null, $branch = null)
     {
         $name = $this->getRemote('name', $name);
@@ -295,6 +375,11 @@ class GitSync extends Git
         return $this->execute("push {$name} {$local_branch}:{$branch}");
     }
 
+    /**
+     * @param string|null $name
+     * @param string|null $branch
+     * @return bool
+     */
     public function sync($name = null, $branch = null)
     {
         $name = $this->getRemote('name', $name);
@@ -310,19 +395,28 @@ class GitSync extends Git
         return true;
     }
 
+    /**
+     * @return string[]
+     */
     public function reset()
     {
-        return $this->execute("reset --hard HEAD");
+        return $this->execute('reset --hard HEAD');
     }
 
+    /**
+     * @return bool
+     */
     public function isWorkingCopyClean()
     {
         $message = 'nothing to commit';
         $output = $this->execute('status');
 
-        return (substr($output[count($output)-1], 0, strlen($message)) === $message);
+        return strpos($output[count($output) - 1], $message) === 0;
     }
 
+    /**
+     * @return bool
+     */
     public function hasChangesToCommit()
     {
         $folders = $this->config['folders'];
@@ -330,19 +424,25 @@ class GitSync extends Git
 
         foreach ($folders as $folder) {
             $folder = explode('/', $folder);
-            $paths[] = is_array($folder) ? array_shift($folder) : $folder;
+            $paths[] = array_shift($folder);
         }
 
         $message = 'nothing to commit';
         $output = $this->execute('status ' . implode(' ', $paths));
 
-        return (substr($output[count($output)-1], 0, strlen($message)) !== $message);
+        return strpos($output[count($output) - 1], $message) !== 0;
     }
 
+    /**
+     * @param string $command
+     * @param bool $quiet
+     * @return string[]
+     */
     public function execute($command, $quiet = false)
     {
         try {
             $bin = Helper::getGitBinary($this->getGitConfig('bin', 'git'));
+            /** @var string $version */
             $version = Helper::isGitInstalled(true);
 
             // -C <path> supported from 1.8.5 and above
@@ -354,17 +454,17 @@ class GitSync extends Git
 
             $command .= ' 2>&1';
 
-            if (DIRECTORY_SEPARATOR == '/') {
+            if (DIRECTORY_SEPARATOR === '/') {
                 $command = 'LC_ALL=C ' . $command;
             }
 
             if ($this->getConfig('logging', false)) {
-                $log_command = Helper::preventReadablePassword($command, $this->password);
+                $log_command = Helper::preventReadablePassword($command, $this->password ?? '');
                 $this->grav['log']->notice('gitsync[command]: ' . $log_command);
 
                 exec($command, $output, $returnValue);
 
-                $log_output = Helper::preventReadablePassword(implode("\n", $output), $this->password);
+                $log_output = Helper::preventReadablePassword(implode("\n", $output), $this->password ?? '');
                 $this->grav['log']->notice('gitsync[output]: ' . $log_output);
             } else {
                 exec($command, $output, $returnValue);
@@ -377,10 +477,10 @@ class GitSync extends Git
             return $output;
         } catch (\RuntimeException $e) {
             $message = $e->getMessage();
-            $message = Helper::preventReadablePassword($message, $this->password);
+            $message = Helper::preventReadablePassword($message, $this->password ?? '');
 
             // handle scary messages
-            if (Utils::contains($message, "remote: error: cannot lock ref")) {
+            if (Utils::contains($message, 'remote: error: cannot lock ref')) {
                 $message = 'GitSync: An error occurred while trying to synchronize. This could mean GitSync is already running. Please try again.';
             }
 
@@ -388,18 +488,33 @@ class GitSync extends Git
         }
     }
 
+    /**
+     * @param string $type
+     * @param mixed $value
+     * @return mixed
+     */
     public function getGitConfig($type, $value)
     {
-        return isset($this->config['git']) && isset($this->config['git'][$type]) ? $this->config['git'][$type] : $value;
+        return $this->config['git'][$type] ?? $value;
     }
 
+    /**
+     * @param string $type
+     * @param mixed $value
+     * @return mixed
+     */
     public function getRemote($type, $value)
     {
-        return !$value && isset($this->config['remote']) ? $this->config['remote'][$type] : $value;
+        return $value ?: ($this->config['remote'][$type] ?? $value);
     }
 
+    /**
+     * @param string $type
+     * @param mixed $value
+     * @return mixed
+     */
     public function getConfig($type, $value)
     {
-        return !$value && isset($this->config[$type]) ? $this->config[$type] : $value;
+        return $value ?: ($this->config[$type] ?? $value);
     }
 }
