@@ -1,9 +1,11 @@
 <?php
 namespace Grav\Plugin\TNTSearch;
 
+use Grav\Common\Config\Config;
 use Grav\Common\Grav;
 use Grav\Common\Yaml;
 use Grav\Common\Page\Page;
+use Grav\Plugin\TNTSearchPlugin;
 
 class GravConnector extends \PDO
 {
@@ -12,7 +14,7 @@ class GravConnector extends \PDO
 
     }
 
-    public function getAttribute($attribute)
+    public function getAttribute($attribute): bool
     {
         return false;
     }
@@ -22,10 +24,11 @@ class GravConnector extends \PDO
         $counter = 0;
         $results = [];
 
+        /** @var Config $config */
         $config = Grav::instance()['config'];
         $filter = $config->get('plugins.tntsearch.filter');
         $default_process = $config->get('plugins.tntsearch.index_page_by_default');
-        $gtnt = \Grav\Plugin\TNTSearchPlugin::getSearchObjectType();
+        $gtnt = TNTSearchPlugin::getSearchObjectType();
 
 
         if ($filter && array_key_exists('items', $filter)) {
@@ -53,17 +56,17 @@ class GravConnector extends \PDO
 
             // Only process what's configured
             if (!$process) {
-                echo("Skipped $counter $route\n");
+                echo("Skipped {$counter} {$route}\n");
                 continue;
             }
 
             try {
                 $fields = $gtnt->indexPageData($page);
                 $results[] = (array) $fields;
-                $display_route = isset($fields->display_route) ? $fields->display_route : $route;
-                echo("Added   $counter $display_route\n");
+                $display_route = $fields->display_route ?? $route;
+                echo("Added   {$counter} {$display_route}\n");
             } catch (\Exception $e) {
-                echo("Skipped $counter $route - {$e->getMessage()}\n");
+                echo("Skipped {$counter} {$route} - {$e->getMessage()}\n");
                 continue;
             }
         }
