@@ -1671,8 +1671,11 @@ class Admin
     {
         $reports = new ArrayCollection();
 
+        /** @var Pages $pages */
+        $pages = $this->grav['pages'];
+
         // Default to XSS Security Report
-        $result = Security::detectXssFromPages($this->grav['pages'], true);
+        $result = Security::detectXssFromPages($pages, true);
 
         $reports['Grav Security Check'] = $this->grav['twig']->processTemplate('reports/security.html.twig', [
             'result' => $result,
