@@ -2,9 +2,12 @@
 
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
+use Grav\Common\Config\Config;
 use Grav\Common\Data\Data;
 use Grav\Common\Grav;
 use Grav\Common\Plugin;
+use Grav\Common\Scheduler\Scheduler;
 use Grav\Plugin\GitSync\AdminController;
 use Grav\Plugin\GitSync\GitSync;
 use Grav\Plugin\GitSync\Helper;
@@ -17,7 +20,9 @@ use RocketTheme\Toolbox\Event\Event;
  */
 class GitSyncPlugin extends Plugin
 {
+    /** @var AdminController|null */
     protected $controller;
+    /** @var GitSync */
     protected $git;
 
     /**
@@ -26,7 +31,10 @@ class GitSyncPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized'   => ['onPluginsInitialized', 1000],
+            'onPluginsInitialized'   => [
+                ['autoload', 100000],
+                ['onPluginsInitialized', 1000]
+            ],
             'onPageInitialized'      => ['onPageInitialized', 0],
             'onFormProcessed'        => ['onFormProcessed', 0],
             'onSchedulerInitialized' => ['onSchedulerInitialized', 0]
@@ -34,11 +42,21 @@ class GitSyncPlugin extends Plugin
     }
 
     /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload() : ClassLoader
+    {
+        return require __DIR__ . '/vendor/autoload.php';
+    }
+
+    /**
      * @return string
      */
     public static function generateWebhookSecret()
     {
-        return bin2hex(openssl_random_pseudo_bytes(24));
+        return static::generateHash(24);
     }
 
     /**
@@ -46,7 +64,7 @@ class GitSyncPlugin extends Plugin
      */
     public static function generateRandomWebhook()
     {
-        return '/_git-sync-' . bin2hex(openssl_random_pseudo_bytes(6));
+        return '/_git-sync-' . static::generateHash(6);
     }
 
     /**
@@ -54,7 +72,6 @@ class GitSyncPlugin extends Plugin
      */
     public function onPluginsInitialized()
     {
-        require_once __DIR__ . '/vendor/autoload.php';
         $this->enable(['gitsync' => ['synchronize', 0]]);
         $this->init();
 
@@ -72,57 +89,62 @@ class GitSyncPlugin extends Plugin
             ]);
 
             return;
-        } else {
-            $config = $this->config->get('plugins.' . $this->name);
-            $route = $this->grav['uri']->route();
-            $webhook = isset($config['webhook']) ? $config['webhook'] : false;
-            $secret = isset($config['webhook_secret']) ? $config['webhook_secret'] : false;
-            $enabled = isset($config['webhook_enabled']) ? $config['webhook_enabled'] : false;
+        }
 
-            if ($route === $webhook && $_SERVER['REQUEST_METHOD'] === 'POST') {
-                if ($secret && $enabled) {
-                    if (!$this->isRequestAuthorized($secret)) {
-                        http_response_code(401);
-                        header('Content-Type: application/json');
-                        echo json_encode([
-                            'status' => 'error',
-                            'message' => 'Unauthorized request'
-                        ]);
-                        exit;
-                    }
-                }
-                try {
-                    $this->synchronize();
-                    header('Content-Type: application/json');
-                    echo json_encode([
-                        'status' => 'success',
-                        'message' => 'GitSync completed the synchronization'
-                    ]);
-                } catch (\Exception $e) {
-                    http_response_code(500);
+        $config = $this->config->get('plugins.' . $this->name);
+        $route = $this->grav['uri']->route();
+        $webhook = $config['webhook'] ?? false;
+        $secret = $config['webhook_secret'] ?? false;
+        $enabled = $config['webhook_enabled'] ?? false;
+
+        if ($route === $webhook && $_SERVER['REQUEST_METHOD'] === 'POST') {
+            if ($secret && $enabled) {
+                if (!$this->isRequestAuthorized($secret)) {
+                    http_response_code(401);
                     header('Content-Type: application/json');
                     echo json_encode([
                         'status' => 'error',
-                        'message' => 'GitSync failed to synchronize'
+                        'message' => 'Unauthorized request'
                     ]);
+                    exit;
                 }
-                exit;
             }
+            try {
+                $this->synchronize();
+                header('Content-Type: application/json');
+                echo json_encode([
+                    'status' => 'success',
+                    'message' => 'GitSync completed the synchronization'
+                ]);
+            } catch (\Exception $e) {
+                http_response_code(500);
+                header('Content-Type: application/json');
+                echo json_encode([
+                    'status' => 'error',
+                    'message' => 'GitSync failed to synchronize'
+                ]);
+            }
+            exit;
         }
     }
 
     /**
      * Returns true if the request contains a valid signature or token
      * @param  string $secret local secret
-     * @return boolean         whether or not the request is authorized
+     * @return bool           whether or not the request is authorized
      */
     public function isRequestAuthorized($secret)
     {
         if (isset($_SERVER['HTTP_X_HUB_SIGNATURE'])) {
-            $payload = file_get_contents('php://input');
+            $payload = file_get_contents('php://input') ?: '';
+
             return $this->isGithubSignatureValid($secret, $_SERVER['HTTP_X_HUB_SIGNATURE'], $payload);
-        } elseif (isset($_SERVER['HTTP_X_GITLAB_TOKEN'])) {
+        }
+        if (isset($_SERVER['HTTP_X_GITLAB_TOKEN'])) {
             return $this->isGitlabTokenValid($secret, $_SERVER['HTTP_X_GITLAB_TOKEN']);
+        } else {
+            $payload = file_get_contents('php://input');
+            return $this->isGiteaSecretValid($secret, $payload);
         }
 
         return false;
@@ -134,28 +156,42 @@ class GitSyncPlugin extends Plugin
      * @param  string $secret The webhook secret
      * @param  string $signatureHeader The signature of the webhook request
      * @param  string $payload The webhook request body
-     * @return boolean                 Whether the signature is valid or not
+     * @return bool            Whether the signature is valid or not
      */
     public function isGithubSignatureValid($secret, $signatureHeader, $payload)
     {
-        list($algorigthm, $signature) = explode('=', $signatureHeader);
+        [$algorithm, $signature] = explode('=', $signatureHeader);
 
-        if ($signature === hash_hmac($algorigthm, $payload, $secret)) {
-            return true;
-        }
-
-        return false;
+        return $signature === hash_hmac($algorithm, $payload, $secret);
     }
 
     /**
      * Returns true if given Gitlab token matches secret
      * @param  string $secret local secret
      * @param  string $token token received from Gitlab webhook request
-     * @return boolean        whether or not secret and token match
+     * @return bool          whether or not secret and token match
      */
     public function isGitlabTokenValid($secret, $token)
     {
         return $secret === $token;
+    }
+
+    /**
+     * Returns true if secret contained in the payload matches the client
+     * secret
+     * @param  string $secret The webhook secret
+     * @param  string $payload The webhook request body
+     * @return boolean Whether the client secret matches the payload secret or
+     * not
+     */
+    public function isGiteaSecretValid($secret, $payload)
+    {
+        $payload = json_decode($payload, true);
+        if (!empty($payload) && isset($payload['secret'])) {
+            return $secret === $payload['secret'];
+        }
+
+        return false;
     }
 
     public function onAdminMenu()
@@ -186,14 +222,15 @@ class GitSyncPlugin extends Plugin
         if ($this->isAdmin()) {
             /** @var AdminController controller */
             $this->controller = new AdminController($this);
+            $this->git = &$this->controller->git;
         } else {
-            $this->controller      = new \stdClass;
-            $this->controller->git = new GitSync($this);
+            $this->git = new GitSync();
         }
-
-        $this->git = $this->controller->git;
     }
 
+    /**
+     * @return bool
+     */
     public function synchronize()
     {
         if (!Helper::isGitInstalled() || !Helper::isGitInitialized()) {
@@ -216,11 +253,9 @@ class GitSyncPlugin extends Plugin
 
     public function onSchedulerInitialized(Event $event)
     {
-
         /** @var Config $config */
         $config = Grav::instance()['config'];
         $run_at = $config->get('plugins.git-sync.sync.cron_at', '0 12,23 * * *');
-
 
         if ($config->get('plugins.git-sync.sync.cron_enable', false)) {
             /** @var Scheduler $scheduler */
@@ -230,6 +265,9 @@ class GitSyncPlugin extends Plugin
         }
     }
 
+    /**
+     * @return bool
+     */
     public function reset()
     {
         if (!Helper::isGitInstalled() || !Helper::isGitInitialized()) {
@@ -255,17 +293,17 @@ class GitSyncPlugin extends Plugin
 
     /**
      * Set needed variables to display cart.
+     *
+     * @return bool
      */
     public function onTwigSiteVariables()
     {
-        // workaround for admin plugin issue that doesn't properly unsubscribe
-        // events upon plugin uninstall
-        if (!class_exists('Grav\Plugin\GitSync\Helper')) {
+        // workaround for admin plugin issue that doesn't properly unsubscribe events upon plugin uninstall
+        if (!class_exists(Helper::class)) {
             return false;
         }
 
         $user = $this->grav['user'];
-
         if (!$user->authenticated) {
             return false;
         }
@@ -293,117 +331,127 @@ class GitSyncPlugin extends Plugin
 
     public function onPageInitialized()
     {
-        if ($this->isAdmin() && $this->controller->isActive()) {
+        if ($this->controller && $this->controller->isActive()) {
             $this->controller->execute();
             $this->controller->redirect();
         }
     }
 
-    public function onAdminSave($event)
+    /**
+     * @param Event $event
+     * @return Data|true
+     */
+    public function onAdminSave(Event $event)
     {
         $obj           = $event['object'];
         $adminPath 	   = trim($this->grav['admin']->base, '/');
-        $isPluginRoute = $this->grav['uri']->path() == "/$adminPath/plugins/" . $this->name;
+        $isPluginRoute = $this->grav['uri']->path() === "/$adminPath/plugins/" . $this->name;
 
         if ($obj instanceof Data) {
             if (!$isPluginRoute || !Helper::isGitInstalled()) {
                 return true;
-            } else {
-                // empty password, keep current one or encrypt if haven't already
-                $password = $obj->get('password', false);
-                if (!$password) { // set to !()
-                    $current_password = $this->controller->git->getPassword();
-                    // password exists but was never encrypted
-                    if (substr($current_password, 0, 8) !== 'gitsync-') {
-                        $current_password = Helper::encrypt($current_password);
-                    }
-                } else {
-                    // password is getting changed
-                    $current_password = Helper::encrypt($password);
-                }
-
-                $obj->set('password', $current_password);
             }
+
+            // empty password, keep current one or encrypt if haven't already
+            $password = $obj->get('password', false);
+            if (!$password) { // set to !()
+                $current_password = $this->git->getPassword();
+                // password exists but was never encrypted
+                if ($current_password && strpos($current_password, 'gitsync-') !== 0) {
+                    $current_password = Helper::encrypt($current_password);
+                }
+            } else {
+                // password is getting changed
+                $current_password = Helper::encrypt($password);
+            }
+
+            $obj->set('password', $current_password);
         }
 
         return $obj;
     }
 
-    public function onAdminAfterSave($event)
+    /**
+     * @param Event $event
+     */
+    public function onAdminAfterSave(Event $event)
     {
         if (!$this->grav['config']->get('plugins.git-sync.sync.on_save', true)) {
-            return true;
+            return;
         }
 
         $obj           = $event['object'];
         $adminPath	   = trim($this->grav['admin']->base, '/');
         $uriPath       = $this->grav['uri']->path();
-        $isPluginRoute = $uriPath == "/$adminPath/plugins/" . $this->name;
+        $isPluginRoute = $uriPath === "/$adminPath/plugins/" . $this->name;
 
         if ($obj instanceof Data) {
-            $folders = $this->controller->git->getConfig('folders', $event['object']->get('folders', []));
+            $folders = $this->git->getConfig('folders', $event['object']->get('folders', []));
             $data_type = preg_replace('#^/' . preg_quote($adminPath, '#') . '/#', '', $uriPath);
             $data_type = explode('/', $data_type);
             $data_type = array_shift($data_type);
 
-            if (!Helper::isGitInstalled() || (!$isPluginRoute && !in_array($this->getFolderMapping($data_type), $folders))) {
-                return true;
+            if (null === $data_type || !Helper::isGitInstalled() || (!$isPluginRoute && !in_array($this->getFolderMapping($data_type), $folders, true))) {
+                return;
             }
 
             if ($isPluginRoute) {
-                $this->controller->git->setConfig($obj);
+                $this->git->setConfig($obj->toArray());
 
                 // initialize git if not done yet
-                $this->controller->git->initializeRepository();
+                $this->git->initializeRepository();
 
                 // set committer and remote data
-                $this->controller->git->setUser();
-                $this->controller->git->addRemote();
+                $this->git->setUser();
+                $this->git->addRemote();
             }
         }
 
         $this->synchronize();
-
-        return true;
     }
 
     public function onAdminAfterSaveAs()
     {
-        if ($this->grav['config']->get('plugins.git-sync.sync.on_save', true)) {
+        if ($this->grav['config']->get('plugins.git-sync.sync.on_save', true))
+        {
             $this->synchronize();
         }
-
-        return true;
     }
 
     public function onAdminAfterDelete()
     {
-        if ($this->grav['config']->get('plugins.git-sync.sync.on_delete', true)) {
+        if ($this->grav['config']->get('plugins.git-sync.sync.on_delete', true))
+        {
             $this->synchronize();
         }
-
-        return true;
     }
 
     public function onAdminAfterMedia()
     {
-        if ($this->grav['config']->get('plugins.git-sync.sync.on_media', true)) {
+        if ($this->grav['config']->get('plugins.git-sync.sync.on_media', true))
+        {
             $this->synchronize();
         }
-
-        return true;
     }
 
+    /**
+     * @param Event $event
+     */
     public function onFormProcessed(Event $event)
     {
         $action = $event['action'];
 
-        if ($action == 'gitsync') {
+        if ($action === 'gitsync') {
             $this->synchronize();
         }
     }
 
-    public function getFolderMapping($data_type) {
+    /**
+     * @param string $data_type
+     * @return string|null
+     */
+    public function getFolderMapping($data_type)
+    {
         switch ($data_type) {
             case 'user':
                 return 'accounts';
@@ -415,5 +463,27 @@ class GitSyncPlugin extends Plugin
             case 'pages':
                 return $data_type;
         }
+
+        return null;
+    }
+
+    /**
+     * @param int $len
+     * @return string
+     */
+    protected static function generateHash(int $len): string
+    {
+        $bytes = openssl_random_pseudo_bytes($len, $isStrong);
+
+        if ($bytes === false) {
+            throw new \RuntimeException('Could not generate hash');
+        }
+
+        if ($isStrong === false) {
+            // It's ok not to be strong [EA].
+            $isStrong = true;
+        }
+
+        return bin2hex($bytes);
     }
 }
