@@ -57,12 +57,6 @@ class Controller
     protected $prefix = 'task';
 
     /**
-     * @var RememberMe\RememberMe
-     * @deprecated 2.0 Use $grav['login']->rememberMe() instead
-     */
-    protected $rememberMe;
-
-    /**
      * @var Login
      */
     protected $login;
@@ -78,8 +72,6 @@ class Controller
         $this->action = $action;
         $this->login = $this->grav['login'];
         $this->post = $post ? $this->getPost($post) : [];
-
-        $this->rememberMe();
     }
 
     /**
@@ -353,7 +345,7 @@ class Controller
             $lang = '';
         }
 
-        $reset_link = $this->grav['base_url_absolute'] . $lang . $this->grav['config']->get('plugins.login.route_reset') . '/task:login.reset/token' . $param_sep . $token . '/user' . $param_sep . $user->username . '/nonce' . $param_sep . Utils::getNonce('reset-form');
+        $reset_link = $this->grav['base_url_absolute'] . $lang . $this->grav['config']->get('plugins.login.route_reset') . '/task' . $param_sep . 'login.reset/token' . $param_sep . $token . '/user' . $param_sep . $user->username . '/nonce' . $param_sep . Utils::getNonce('reset-form');
 
         $sitename = $this->grav['config']->get('site.title', 'Website');
 
@@ -554,9 +546,7 @@ class Controller
      */
     public function rememberMe($var = null)
     {
-        $this->rememberMe = $this->login->rememberMe($var);
-
-        return $this->rememberMe;
+        return $this->login->rememberMe($var);
     }
 
     /**
