@@ -1,7 +1,6 @@
 <?php
 namespace Grav\Plugin\Console;
 
-use Grav\Common\Grav;
 use Grav\Console\ConsoleCommand;
 use Grav\Plugin\TNTSearchPlugin;
 use Symfony\Component\Console\Input\InputOption;
@@ -37,9 +36,9 @@ class IndexerCommand extends ConsoleCommand
     protected function configure()
     {
         $this
-            ->setName("index")
-            ->addOption("alt", null, InputOption::VALUE_NONE, 'alternative output')
-            ->setDescription("TNTSearch Indexer")
+            ->setName('index')
+            ->addOption('alt', null, InputOption::VALUE_NONE, 'alternative output')
+            ->setDescription('TNTSearch Indexer')
             ->setHelp('The <info>index command</info> re-indexes the search engine');
     }
 
@@ -48,6 +47,8 @@ class IndexerCommand extends ConsoleCommand
      */
     protected function serve()
     {
+        $this->initializePages();
+
         $alt_output = $this->input->getOption('alt') ? true : false;
 
         if ($alt_output) {
@@ -68,11 +69,7 @@ class IndexerCommand extends ConsoleCommand
     {
         error_reporting(1);
 
-        $grav = Grav::instance();
-        $grav->fireEvent('onPluginsInitialized');
-        $grav->fireEvent('onThemeInitialized');
-
-        list($status, $msg, $output) = TNTSearchPlugin::indexJob();
+        [$status, $msg, $output] = TNTSearchPlugin::indexJob();
 
         $this->output->write($output);
         $this->output->writeln('');
