@@ -1,9 +1,7 @@
 <?php
 namespace Grav\Plugin\Console;
 
-use Grav\Common\Grav;
 use Grav\Console\ConsoleCommand;
-//use Grav\Plugin\TNTSearch\GravIndexer;
 use Grav\Plugin\TNTSearchPlugin;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
@@ -42,8 +40,8 @@ class QueryCommand extends ConsoleCommand
     protected function configure()
     {
         $this
-            ->setName("query")
-            ->setDescription("TNTSearch Query")
+            ->setName('query')
+            ->setDescription('TNTSearch Query')
             ->addArgument(
                 'query',
                 InputArgument::REQUIRED,
@@ -64,36 +62,17 @@ class QueryCommand extends ConsoleCommand
      */
     protected function serve()
     {
+        $this->setLanguage($this->input->getOption('language'));
+        $this->initializePages();
+
         $this->doQuery();
         $this->output->writeln('');
     }
 
     private function doQuery()
     {
-        $grav = Grav::instance();
-
-
-        // Initialize Plugins
-        $grav->fireEvent('onPluginsInitialized');
-
-        // Set Language if one passed in
-        $language = $grav['language'];
-        if ($language->enabled()) {
-            $lang = $this->input->getOption('language');
-            if ($lang && $language->validate($lang)) {
-                $language->setActive($lang);
-            } else {
-                $language->setActive($language->getDefault());
-            }
-        }
-
-        $grav['debugger']->enabled(false);
-        $grav['twig']->init();
-        $grav['pages']->init();
-
         $gtnt = TNTSearchPlugin::getSearchObjectType(['json' => true]);
         print_r($gtnt->search($this->input->getArgument('query')));
     }
-
 }
 
